change uint32_t to size_t

+++ Game/CharacterOcto.cpp
@@ -84,28 +84,28 @@ void	CharacterOcto::setup(ABiome & biome)
 	m_waterLevel = biome.getWaterLevel();
 	m_box->setGameObject(this);
 	m_box->setSize(sf::Vector2f(30.f, 85.f));
-	m_box->setCollisionType(static_cast<std::uint32_t>(GameObjectType::Player));
-	std::uint32_t mask = static_cast<std::uint32_t>(GameObjectType::Portal)
-		| static_cast<std::uint32_t>(GameObjectType::GroundTransformNanoRobot)
-		| static_cast<std::uint32_t>(GameObjectType::RepairNanoRobot)
-		| static_cast<std::uint32_t>(GameObjectType::JumpNanoRobot)
-		| static_cast<std::uint32_t>(GameObjectType::Elevator)
-		| static_cast<std::uint32_t>(GameObjectType::Tent)
-		| static_cast<std::uint32_t>(GameObjectType::Bouibouik);
+	m_box->setCollisionType(static_cast<std::size_t>(GameObjectType::Player));
+	std::size_t mask = static_cast<std::size_t>(GameObjectType::Portal)
+		| static_cast<std::size_t>(GameObjectType::GroundTransformNanoRobot)
+		| static_cast<std::size_t>(GameObjectType::RepairNanoRobot)
+		| static_cast<std::size_t>(GameObjectType::JumpNanoRobot)
+		| static_cast<std::size_t>(GameObjectType::Elevator)
+		| static_cast<std::size_t>(GameObjectType::Tent)
+		| static_cast<std::size_t>(GameObjectType::Bouibouik);
 	m_box->setCollisionMask(mask);
 
 	m_octoEvent.m_octo = this;
 	m_eventBox->setGameObject(&m_octoEvent);
 	m_eventBox->setRadius(400.f);
-	m_eventBox->setCollisionType(static_cast<std::uint32_t>(GameObjectType::PlayerEvent));
-	std::uint32_t maskEvent = static_cast<std::uint32_t>(GameObjectType::Portal)
-		| static_cast<std::uint32_t>(GameObjectType::Elevator)
-		| static_cast<std::uint32_t>(GameObjectType::CedricNpc)
-		| static_cast<std::uint32_t>(GameObjectType::FannyNpc)
-		| static_cast<std::uint32_t>(GameObjectType::FranfranNpc)
-		| static_cast<std::uint32_t>(GameObjectType::GuiNpc)
-		| static_cast<std::uint32_t>(GameObjectType::JuNpc)
-		| static_cast<std::uint32_t>(GameObjectType::TurbanNpc);
+	m_eventBox->setCollisionType(static_cast<std::size_t>(GameObjectType::PlayerEvent));
+	std::size_t maskEvent = static_cast<std::size_t>(GameObjectType::Portal)
+		| static_cast<std::size_t>(GameObjectType::Elevator)
+		| static_cast<std::size_t>(GameObjectType::CedricNpc)
+		| static_cast<std::size_t>(GameObjectType::FannyNpc)
+		| static_cast<std::size_t>(GameObjectType::FranfranNpc)
+		| static_cast<std::size_t>(GameObjectType::GuiNpc)
+		| static_cast<std::size_t>(GameObjectType::JuNpc)
+		| static_cast<std::size_t>(GameObjectType::TurbanNpc);
 	m_eventBox->setCollisionMask(maskEvent);
 	m_eventBox->setApplyGravity(false);
 	m_eventBox->setType(AShape::Type::e_trigger);
